+++ app/webpage/src/index.tsx
@@ -1,7 +1,6 @@
 import { Root } from './components/Root';
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import { Header } from './components/Header';
 
 export default function App() {
     return (
