+++ app/webpage/src/types/NetState.ts
@@ -1,0 +1,1 @@
+export type NetState = "none" | "offering" | "answering" | "connected";
